Issue#69 - Fix
Flag "can_be_edited" defined more precisely by by testing on
"input_is_open" status of a training and on the fact that the user is
RESPENS of the structure. See the following file :
    - mecc/apps/utils/queries.py - line 80

Flag "input_is_open" removed from the template of the mecctable. Flag
"can_edit" defined more precisely by testing on input_is_open and on the
fact that the user is poweruser on the training. See the followin files:
    - mecc/apps/mecctable/views.py
    - mec/templates/mecctable/tree_view.html